RavenDB-13110 - Outgoing Replication Handler abstraction
